refactor(workspace): split e2e workspace hosts
Move the fake runtime workspace host behavior into dedicated E2E modules so the web and desktop hosts stay focused on their real implementations and only select a strategy once.
